feat(service): Implement core business logic layer
Implements the service layer for products, users, carts, and orders, establishing the core business logic for the application.

- Creates service interfaces (`ProductService`, `UserService`, `CartService`, `OrderService`) to define business contracts.
- Adds full implementations (`...Impl`) for each service, handling logic for registration, cart management, and order placement.
- Introduces DTOs for API data transfer (`UserRegistrationDTO`, `CartDTO`, etc.).
- Adds a `PasswordEncoder` bean for security.
- Defines custom exceptions for handling business-specific errors.

fix(model): Use BigDecimal for currency fields

Refactors the Product and Order entities to use `BigDecimal` instead of `Double` for all price and currency fields.

This is a critical fix to prevent floating-point inaccuracies in all financial calculations, ensuring data integrity for prices and totals.

fix(repository): Correct method signatures for pagination

Updates repository interfaces to return `Page<T>` instead of `List<T>` for methods that accept a `Pageable` parameter.

- Modifies `ProductRepository.findByNameContainingIgnoreCase`
- Modifies `OrderRepository.findByUser`

This resolves a type mismatch in the service layer and correctly implements pagination support at the data access level.

+++ src/main/java/com/muiyurocodes/ecommerc/model/OrderItem.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @Setter
@@ -26,6 +28,6 @@ public class OrderItem {
 
     private int quantity;
 
-    private Double price; // Price at the time of purchase
+    private BigDecimal price; // Price at the time of purchase
 
 }
